put the raw data in a dialog
